Move social links to own component and create test
